Convert audit log sequence to 64-bit

+++ netprofile_core/netprofile_core/models.py
@@ -6022,7 +6022,7 @@ class LogData(Base):
 	)
 	id = Column(
 		'logid',
-		UInt32(),
+		UInt64(),
 		Sequence('logs_data_logid_seq'),
 		Comment('Log entry ID'),
 		primary_key=True,
